update pizza, bucket to new api, cleaning code

+++ Pizzeria/src/main/java/com/xzteam/pizzeria/rest/BucketController.java
@@ -1,15 +1,16 @@
 package com.xzteam.pizzeria.rest;
 
-import com.xzteam.pizzeria.api.GenericReply;
 import com.xzteam.pizzeria.api.bucket.BucketApiAddRequest;
 import com.xzteam.pizzeria.api.bucket.BucketApiListReply;
 import com.xzteam.pizzeria.domain.Bucket;
 import com.xzteam.pizzeria.mappers.BucketMapper;
+import com.xzteam.pizzeria.rest.exceptions.NotFoundException;
 import com.xzteam.pizzeria.services.BucketService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
@@ -18,9 +19,9 @@ public class BucketController {
     private static final Logger log = Logger.getLogger(BucketController.class.getName());
 
     @Autowired
-    BucketService bucketService;
+    private BucketService bucketService;
     @Autowired
-    BucketMapper bucketMapper;
+    private BucketMapper bucketMapper;
 
     @RequestMapping(path = "/buckets", method = RequestMethod.GET,
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -28,7 +29,7 @@ public class BucketController {
         BucketApiListReply bucketList = new BucketApiListReply();
         bucketList.buckets.addAll(bucketService.getAll()
                 .stream()
-                .map(bucket -> bucketMapper.toApi(bucket))
+                .map(bucket -> bucketMapper.toApiGet(bucket))
                 .collect(Collectors.toList()));
         return bucketList;
     }
@@ -37,7 +38,7 @@ public class BucketController {
             produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public BucketApiListReply getBucketById(@PathVariable Long id) {
         BucketApiListReply bucketById = new BucketApiListReply();
-        bucketById.buckets.add(bucketMapper.toApi(bucketService.getBucketById(id)));
+        bucketById.buckets.add(bucketMapper.toApiGet(bucketService.getBucketById(id)));
         return bucketById;
     }
 
@@ -47,38 +48,58 @@ public class BucketController {
         BucketApiListReply bucketList = new BucketApiListReply();
         bucketList.buckets.addAll(bucketService.getAllBucketsByClientId(id)
                 .stream()
-                .map(bucket -> bucketMapper.toApi(bucket))
+                .map(bucket -> bucketMapper.toApiGet(bucket))
                 .collect(Collectors.toList()));
         return bucketList;
     }
 
-    @RequestMapping(path = "/buckets/{id}", method = RequestMethod.DELETE,
-            produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public GenericReply delBucket(@PathVariable Long id) {
-        GenericReply reply = new GenericReply();
+    @RequestMapping(path = "/buckets", method = RequestMethod.POST,
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE, consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public BucketApiListReply addBucket(@RequestBody BucketApiAddRequest req) {
+        BucketApiListReply reply = new BucketApiListReply();
         try {
-            bucketService.deleteBucket(id);
+            Bucket bucket = bucketService.addBucket(bucketMapper.fromApiPost(req));
+            reply.buckets.add(bucketMapper.toApiGet(bucket));
         } catch (Exception e) {
-            reply.code = -1;
-            reply.message = e.getMessage();
-            log.warning("Error deleting bucket: " + e.getMessage());
+            log.warning("Error adding bucket: " + e.getMessage());
+            throw e;
         }
         return reply;
     }
 
-    @RequestMapping(path = "/buckets", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_UTF8_VALUE, consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public BucketApiListReply addBucket(@RequestBody BucketApiAddRequest req) {
+    @RequestMapping(path = "/buckets/{id}", method = RequestMethod.PUT,
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE, consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public BucketApiListReply updateBucket(@PathVariable Long id, @Valid @RequestBody BucketApiAddRequest req) {
+        if (!bucketService.exists(id)) {
+            String msg = "Bucket with id=" + id + " not found!";
+            log.warning("Error updating bucket: " + msg);
+            throw new NotFoundException(msg);
+        }
         BucketApiListReply reply = new BucketApiListReply();
         try {
-            Bucket bucket = bucketService.addBucket(bucketMapper.fromApi(req));
-            reply.buckets.add(bucketMapper.toApi(bucket));
+            Bucket bucket = bucketService.updateBucket(bucketMapper.fromApiPut(req, id));
+            reply.buckets.add(bucketMapper.toApiGet(bucket));
         } catch (Exception e) {
-            String msg = "Error adding bucket: " + e.getMessage();
-            reply.code = -1;
-            reply.message = msg;
-            log.warning(msg);
-    }
+            log.warning("Error updating bucket : " + e.getMessage());
+            throw e;
+        }
         return reply;
+    }
+
+    @RequestMapping(path = "/buckets/{id}", method = RequestMethod.DELETE,
+            produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public void deleteBucket(@PathVariable Long id) {
+        if (!bucketService.exists(id)) {
+            String msg = "Bucket with id=" + id + " not found!";
+            log.warning("Error deleting bucket: " + msg);
+            throw new NotFoundException(msg);
+        }
+        try {
+            bucketService.deleteBucket(id);
+        } catch (Exception e) {
+            log.warning("Error deleting bucket : " + e.getMessage());
+            throw e;
+        }
     }
 
 }
